// src/core/interfaces/Role.ts
export enum Staff_Role_Permissions {
  BACK_OFFICE = "BACK_OFFICE",
  TILL = "TILL",
  ADMIN_ACCESS_ON_TILL = "ADMIN_ACCESS_ON_TILL",
  TILL_SETTINGS = "TILL_SETTINGS",
  QUICK_ADD_SETTINGS = "QUICK_ADD_SETTINGS",
  CLOCK_IN_CLOCK_OUT_INFO = "CLOCK_IN_CLOCK_OUT_INFO",
  MANAGER_OVERRIDE = "MANAGER_OVERRIDE",
  NO_SALES = "NO_SALES",
  PETTY_CASH = "PETTY_CASH",
  FLOAT_ADJUSTMENT = "FLOAT_ADJUSTMENT",
  STOCK_SEND = "STOCK_SEND",
  STOCK_RECEIVE = "STOCK_RECEIVE",
  STOCK_TAKE = "STOCK_TAKE",
  PAYOUTS = "PAYOUTS",
  HOLD = "HOLD",
  CLOSE_TILL = "CLOSE_TILL",
  BLIND_END_OF_DAY = "BLIND_END_OF_DAY",
  VOID_ANY_ITEM = "VOID_ANY_ITEM",
  DELETE_UNORDERED_ITEMS = "DELETE_UNORDERED_ITEMS",
  CLEAR_TRANSACTION = "CLEAR_TRANSACTION",
  REMOVE_FROM_TABLE = "REMOVE_FROM_TABLE",
  ITEM_DISCOUNT = "ITEM_DISCOUNT",
  ITEM_DISCOUNT_LIMIT = "ITEM_DISCOUNT_LIMIT",
  ITEM_DISCOUNT_LIMIT_PERCENTAGE = "ITEM_DISCOUNT_LIMIT_PERCENTAGE",
  BASKET_DISCOUNT = "BASKET_DISCOUNT",
}

export interface Role {
  id: string;
  name: string;
  description?: string;
  permissions: Staff_Role_Permissions[];
  createdAt?: string;
  updatedAt?: string;
}